backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "session-trading-backend"
version = "0.1.0"
description = "FastAPI backend for session trading"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "greenlet>=3.2.0,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "websockets>=15.0.0,<16.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
